consolidated types cleanup
