+++ apps/backend/src/services/webhook.service.ts
@@ -2,17 +2,80 @@ import { env } from "cloudflare:workers"
 import type { WebhookQueueMessage } from "@alchemy.run"
 import type { Queue } from "@cloudflare/workers-types"
 import type { Address, Hash } from "viem"
+import type {
+  OrderType,
+  SubscriptionStatus,
+} from "@/constants/subscription.constants"
 import { ErrorCode, HTTPError } from "@/errors/http.errors"
 import { logger } from "@/lib/logger"
 import { WebhookRepository } from "@/repositories/webhook.repository"
 import type { ActivationResult } from "@/services/subscription.service"
-import {
-  type OrderType,
-  type SubscriptionStatus,
-  type SubscriptionUpdatedEventData,
-  WEBHOOK_EVENT_TYPE,
-  type WebhookEvent,
-} from "@/types/webhook.types"
+
+/**
+ * Webhook event types for v1
+ * Following spec: "Single event type with domain-aligned structure"
+ */
+export const WEBHOOK_EVENT_TYPE = "subscription.updated" as const
+
+/**
+ * Order status in webhook events (subset of OrderStatus for external API)
+ */
+export type WebhookOrderStatus = "paid" | "failed"
+
+/**
+ * Subscription data in webhook event
+ */
+export interface WebhookSubscriptionData {
+  id: Hash
+  status: SubscriptionStatus
+  current_period_end?: number // Unix timestamp, present if active
+}
+
+/**
+ * Order data in webhook event (present if event relates to a payment)
+ */
+export interface WebhookOrderData {
+  number: number // Sequential number relative to subscription
+  type: OrderType
+  amount: string
+  status: WebhookOrderStatus
+}
+
+/**
+ * Transaction data in webhook event (present if payment was successful)
+ */
+export interface WebhookTransactionData {
+  hash: Hash
+  amount: string
+  processed_at: number // Unix timestamp
+}
+
+/**
+ * Error data in webhook event (present if payment failed)
+ */
+export interface WebhookErrorData {
+  code: string
+  message: string
+}
+
+/**
+ * Event data structure for subscription.updated
+ */
+export interface SubscriptionUpdatedEventData {
+  subscription: WebhookSubscriptionData
+  order?: WebhookOrderData
+  transaction?: WebhookTransactionData
+  error?: WebhookErrorData
+}
+
+/**
+ * Complete webhook event structure
+ */
+export interface WebhookEvent {
+  type: typeof WEBHOOK_EVENT_TYPE
+  created_at: number // Unix timestamp
+  data: SubscriptionUpdatedEventData
+}
 
 export interface SetWebhookParams {
   accountAddress: Address
